requirements.txt
django==4.0.*
django-admin-sortable2==2.1.*
django-tinymce==3.6.*
environs==10.0.*
pillow==10.1.*
requests==2.31.*
